Use chmod/chgrp inside each RUN command that impacts /home/user
To avoid a massive layer towards the end of the image build, run

  chgrp -R 0 /home/user && chmod -R g=u /home/user

after each command that impacts the /home/user directory. This avoids
creating a layer in one of the last steps, where updating
ownership/permissions on ~2GB of files in /home/user results in a layer
that duplicates all 2GB of those files.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>